requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
beautifulsoup4
sentence-transformers==2.2.2
torch
transformers
